add: checkout chain

+++ Grad_01/BusinessObjects/DTO/PaymentDTOs.cs
@@ -5,9 +5,9 @@ namespace BusinessObjects.DTO
     {
         public string PaymentContent { get; set; } = string.Empty;
         public string PaymentCurrency { get; set; } = "VND";
-        public decimal? RequiredAmount { get; set; }
+        public int? RequiredAmount { get; set; }
         //public string? PaymentLanguage { get; set; } = "vn";
-        //public string? MerchantId { get; set; } = string.Empty;
+        //public string? MerchantId { get; set; } = string.Esmpty;
         //public string? PaymentDestinationId { get; set; } = string.Empty;
         //public string? Signature { get; set; } = string.Empty;
     }
